requirements.txt
Pillow
colorama
customtkinter
pystray
pyserial
winsdk